refactored AOTAutograd code a bit
